Add instructions to replicate experiments

+++ lib/representations.py
@@ -37,7 +37,7 @@ class ModelFactory:
         return BitsAndBytesConfig(
             load_in_4bit=True,
             bnb_4bit_quant_type="nf4",
-            bnb_4bit_compute_dtype=torch.float16,
+            bnb_4bit_compute_dtype=torch.float32,
         )
 
     @staticmethod
